+++ src/main/java/org/example/repositories/MovieRepository.java
@@ -3,4 +3,4 @@ import org.example.entities.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Integer> {}
+public interface MovieRepository extends JpaRepository<Movie, Long> {}
