+++ src/main/java/org/example/repositories/ReviewRepository.java
@@ -3,4 +3,4 @@ import org.example.entities.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Integer> {}
+public interface ReviewRepository extends JpaRepository<Review, Long> {}
